refacto network availability

+++ svelte/kredeum-nfts.svelte
@@ -46,12 +46,13 @@
     >
   </nav>
 
+  <!-- {#key address} -->
   <main class="main-krd" role="main">
     <div id="kredeum-list">
       <section class="content">
         <header>
           <h1>My NFT wallet</h1>
-          {#if network?.create}
+          {#if address && network?.create}
             <a href="#create" class="btn btn-default" title="Create"
               ><i class="fas fa-plus fa-left" />Create</a
             >
@@ -61,9 +62,9 @@
             <Metamask autoconnect="off" bind:address bind:chainId bind:signer />
 
             <div class="col col-xs-12 col-sm-3">
-              <span class="label"
-                >Collection &nbsp;&nbsp;&nbsp;
-                {#if collection}
+              {#if address && collection}
+                <span class="label"
+                  >Collection &nbsp;&nbsp;&nbsp;
                   <a
                     class="info-button"
                     href={explorerCollectionUrl(chainId, collection?.address)}
@@ -73,21 +74,23 @@
                     )}"
                     target="_blank"><i class="fas fa-info-circle" /></a
                   >
-                {/if}
-              </span>
-              <KredeumSelectCollection bind:address bind:chainId bind:collection />
+                </span>
+                <KredeumSelectCollection bind:address bind:chainId bind:collection />
+              {/if}
             </div>
 
             <div class="col col-sm-3">
-              <button
-                class="clear"
-                on:click={() => {
-                  localStorage.clear();
-                  refreshNFTs();
-                }}
-              >
-                <i class="fas fa-redo-alt {refreshing ? 'refresh' : ''}" />
-              </button>
+              {#if address}
+                <button
+                  class="clear"
+                  on:click={() => {
+                    localStorage.clear();
+                    refreshNFTs();
+                  }}
+                >
+                  <i class="fas fa-redo-alt {refreshing ? 'refresh' : ''}" />
+                </button>
+              {/if}
             </div>
           </div>
         </header>
@@ -102,6 +105,7 @@
       </section>
     </div>
   </main>
+  <!-- {/key} -->
 
   <!-- Modal create -->
   <div id="create" class="modal-window">
